requirements/development.txt
-r base.txt
pre-commit==4.2.0
ruff==0.11.8
black==25.1.0
django-debug-toolbar==5.2.0
